feat: MyProfileImage 컴포넌트 구현
